Remove ReadOnlyApplicantProgramService interface
